api/requirements.txt
django-rest-framework==0.1.0
pip-chill==1.0.1
